requirements.txt
langchain==0.3.7
langchain_community==0.3.5
langchain_openai==0.2.6
langchain_pinecone==0.2.0
langchain_qdrant==0.2.0
openai==1.54.3
pinecone==5.3.1
PyPDF2==3.0.1
python-dotenv==1.0.1
qdrant_client==1.12.1
streamlit==1.37.0
tenacity==8.2.3
tqdm==4.66.4
